takes error handling to a different place
