+++ future_webpage/dgf/views.py
@@ -1,7 +1,7 @@
 import random
 
 from django.urls import reverse
-from django.views import generic, defaults
+from django.views import generic
 from django.views.generic import CreateView
 
 from .formsets import FavoriteCourseFormset, HighlightFormset, DiscFormset, AceFormset, VideoFormset
@@ -90,19 +90,3 @@ class MediaIndex(generic.ListView):
         all_videos = list(set(Video.objects.all().values_list('url', flat=True)))
         random.shuffle(all_videos)
         return all_videos
-
-
-def nice_format(dictionary):
-    return '\n'.join(['* **{}**: {}'.format(key, value) for key, value in dictionary.items()])
-
-
-def server_error(request):
-    headers = nice_format(request.headers)
-    body = nice_format(request.POST) if request.method == 'POST' else ' -- '
-    friend = request.user.friend if not request.user.is_anonymous else None
-    Feedback.objects.create(title='Server error on {} {}'.format(request.method, request.get_full_path()),
-                            feedback='# Headers\n{}\n'
-                                     '# Body\n{}'.format(headers, body),
-                            friend=friend)
-
-    return defaults.server_error(request)
